requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
torchvision==0.13.1+cu113 
matplotlib
scikit-image
pytorch-lightning==1.9.3 
diffusers
einops
kornia
tk
Pillow
libtmux